+++ app/api/auth/[...nextauth]/route.ts
@@ -1,76 +1,76 @@
-// import { PrismaAdapter } from "@next-auth/prisma-adapter";
-// import NextAuth, { AuthOptions } from "next-auth";
-// import GithubProvider from "next-auth/providers/github";
-// import GoogleProvider from "next-auth/providers/google";
-// import CredentialsProvider from "next-auth/providers/credentials";
-// import bcrypt from "bcrypt";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import NextAuth, { AuthOptions } from "next-auth";
+import GithubProvider from "next-auth/providers/github";
+import GoogleProvider from "next-auth/providers/google";
+import CredentialsProvider from "next-auth/providers/credentials";
+import bcrypt from "bcrypt";
 
-import option from "@/app/option";
-import NextAuth from 'next-auth';
-
-
-// import prisma from "@/app/libs/prismadb";
+// import option from "@/app/option";
 
 
-// export const authOptions: AuthOptions = {
-//   adapter: PrismaAdapter(prisma),
-//   providers: [
-//     GithubProvider({
-//       clientId: process.env.GITHUB_ID as string,
-//       clientSecret: process.env.GITHUB_SECRET as string,
-//     }),
-//     GoogleProvider({
-//       clientId: process.env.GOOGLE_ID as string,
-//       clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-//     }),
-//     CredentialsProvider({
-//         name: 'credentials',
-//         credentials: {
-//           email: { label: 'email', type: 'text' },
-//           password: { label: 'password', type: 'password' }
-//         },
-//         async authorize(credentials) {
-//           if (!credentials?.email || !credentials?.password) {
-//             throw new Error('Invalid credentials');
-//           }
+
+import prisma from "@/app/libs/prismadb";
+
+
+export const authOptions: AuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  providers: [
+    GithubProvider({
+      clientId: process.env.GITHUB_ID as string,
+      clientSecret: process.env.GITHUB_SECRET as string,
+    }),
+    GoogleProvider({
+      clientId: process.env.GOOGLE_ID as string,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+    }),
+    CredentialsProvider({
+        name: 'credentials',
+        credentials: {
+          email: { label: 'email', type: 'text' },
+          password: { label: 'password', type: 'password' }
+        },
+        async authorize(credentials) {
+          if (!credentials?.email || !credentials?.password) {
+            throw new Error('Invalid credentials');
+          }
   
-//           const user = await prisma.user.findUnique({
-//             where: {
-//               email: credentials.email
-//             }
-//           });
+          const user = await prisma.user.findUnique({
+            where: {
+              email: credentials.email
+            }
+          });
   
-//           if (!user || !user?.hashedPassword) {
-//             throw new Error('Invalid credentials');
-//           }
+          if (!user || !user?.hashedPassword) {
+            throw new Error('Invalid credentials');
+          }
   
-//           const isCorrectPassword = await bcrypt.compare(
-//             credentials.password,
-//             user.hashedPassword
-//           );
+          const isCorrectPassword = await bcrypt.compare(
+            credentials.password,
+            user.hashedPassword
+          );
   
-//           if (!isCorrectPassword) {
-//             throw new Error('Invalid credentials');
-//           }
+          if (!isCorrectPassword) {
+            throw new Error('Invalid credentials');
+          }
   
-//           return user;
-//         }
-//       })
-//   ],
-//   pages: {
-//     signIn: '/'
-//   },
-//   debug: process.env.NODE_ENV === 'development',
-//   session: {
-//     strategy: "jwt"
-//   },
-//   secret: process.env.NEXTAUTH_SECRET,
-// };
+          return user;
+        }
+      })
+  ],
+  pages: {
+    signIn: '/'
+  },
+  debug: process.env.NODE_ENV === 'development',
+  session: {
+    strategy: "jwt"
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+};
 
-// export default NextAuth(authOptions)
+export default NextAuth(authOptions)
 
 
-const handler = NextAuth(option);
+const handler = NextAuth(authOptions);
 
 export { handler as GET, handler as POST}
